Struct conserve::block::BlockWriter

pub struct BlockWriter {
    // some fields omitted
}

Write body data to a data block, compressed, and stored by its hash.

A BlockWriter is a single-use object that writes a single block.

Data is compressed and its hash is accumulated until writing is complete.

TODO: Implement all of std::io::Write?

Methods

impl BlockWriter

fn new() -> BlockWriter

Make a new BlockWriter, to write one block into a block data directory dir.

fn write_all(self: &mut BlockWriter, buf: &[u8]) -> Result<()>

Write all the contents of buf into this block.

If this returns an error then it's possible that the block was partly written, and the caller should discard it.

fn finish(self: BlockWriter) -> Result<(Vec<u8>, BlockHash)>

Finish compression, and return the compressed bytes and a hex hash.

Callers normally want BlockDir.store instead, which will finish and consume the writer.